chore: use single Withdraw component for mobile and desktop versions (EX-138)
https://bitzlato.atlassian.net/browse/EX-138
