+++ web/src/components/Label/Label.postcss
@@ -1,9 +1,14 @@
+.smSize {
+  font-size: var(--font-size-smaller);
+}
+
 .lgSize {
   font-size: var(--font-size-bigger);
 }
 
-.smSize {
-  font-size: var(--font-size-smaller);
+.xlSize {
+  font-size: var(--font-size-xl);
+  font-weight: 500;
 }
 
 .primaryColor {
@@ -39,11 +44,9 @@
 }
 
 .ellipsis {
-  flex: 1;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
-  width: 0;
 }
 
 .center {
