+++ web/src/components/NumberInput/NumberInput.postcss
@@ -23,6 +23,7 @@
     background: var(--body-background-color) !important;
     left: 12px !important;
     margin-top: -7px !important;
+    max-width: 85%;
     padding: 0 calc(var(--gap) * 0.7);
     position: absolute;
     top: 0 !important;
