+++ web/src/index.pcss
@@ -8,6 +8,7 @@
   --font-size: 12px !important;
   --font-size-smaller: 11px;
   --font-size-bigger: 14px;
+  --font-size-xl: 18px;
   --line-height: 1.3;
 
   /* layout */
@@ -190,6 +191,7 @@ body {
     --font-size: 14px !important;
     --font-size-smaller: 12px;
     --font-size-bigger: 16px;
+    --font-size-xl: 18px;
   }
 }
