+++ web/src/mobile/components/index.ts
@@ -17,4 +17,3 @@ export * from './WalletDepositBody';
 export * from './WalletHeader';
 export * from './WalletItem';
 export * from './WalletsButtons';
-export * from './WalletWithdrawBody';
