+++ web/src/screens/WalletsScreen/WalletsScreen.pcss
@@ -404,8 +404,6 @@
 }
 
 .pg-history-elem__wallet {
-  padding: calc(var(--gap) * 3);
-
   .pg-history-elem__label {
     color: var(--wallets-balance-color);
     font-size: calc(var(--gap) * 5);
